+++ sdwan_policy_objects.tf
@@ -34,12 +34,22 @@ resource "sdwan_policy_object_as_path_list" "policy_object_as_path_list" {
 }
 
 resource "sdwan_policy_object_class_map" "policy_object_class_map" {
-  for_each           = { for p in try(local.feature_profiles.policy_object_profile.class_maps, {}) : p.name => p }
+  for_each           = { for p in try(local.feature_profiles.policy_object_profile.forwarding_classes, {}) : p.name => p }
   name               = each.value.name
   description        = try(each.value.description, null)
   feature_profile_id = sdwan_policy_object_feature_profile.policy_object_feature_profile[0].id
   entries = [{
     queue = each.value.queue
+  }]
+}
+
+resource "sdwan_policy_object_color_list" "policy_object_color_list" {
+  for_each           = { for p in try(local.feature_profiles.policy_object_profile.color_lists, {}) : p.name => p }
+  name               = each.value.name
+  description        = null # not supported in the UI
+  feature_profile_id = sdwan_policy_object_feature_profile.policy_object_feature_profile[0].id
+  entries = [for e in try(each.value.colors, []) : {
+    color = e
   }]
 }
 
@@ -132,6 +142,96 @@ resource "sdwan_policy_object_policer" "policy_object_policer" {
   }]
 }
 
+resource "sdwan_policy_object_preferred_color_group" "policy_object_preferred_color_group" {
+  for_each           = { for p in try(local.feature_profiles.policy_object_profile.preferred_color_groups, {}) : p.name => p }
+  name               = each.value.name
+  description        = try(each.value.description, null)
+  feature_profile_id = sdwan_policy_object_feature_profile.policy_object_feature_profile[0].id
+  entries = [{
+    primary_color_preference   = each.value.primary_colors
+    primary_path_preference    = try(each.value.primary_path_preference, null)
+    secondary_color_preference = try(each.value.secondary_colors, null)
+    secondary_path_preference  = try(each.value.secondary_path_preference, null)
+    tertiary_color_preference  = try(each.value.tertiary_colors, null)
+    tertiary_path_preference   = try(each.value.tertiary_path_preference, null)
+  }]
+}
+
+resource "sdwan_policy_object_security_data_ipv4_prefix_list" "policy_object_security_data_ipv4_prefix_list" {
+  for_each           = { for p in try(local.feature_profiles.policy_object_profile.security_data_ipv4_prefix_lists, {}) : p.name => p }
+  name               = each.value.name
+  description        = null # not supported in the UI
+  feature_profile_id = sdwan_policy_object_feature_profile.policy_object_feature_profile[0].id
+  entries = [for e in try(each.value.prefixes, []) : {
+    ip_prefix          = try(e, null)
+    ip_prefix_variable = null # not supported in the UI
+  }]
+}
+
+resource "sdwan_policy_object_security_fqdn_list" "policy_object_security_fqdn_list" {
+  for_each           = { for p in try(local.feature_profiles.policy_object_profile.security_fqdn_lists, {}) : p.name => p }
+  name               = each.value.name
+  feature_profile_id = sdwan_policy_object_feature_profile.policy_object_feature_profile[0].id
+  entries = [for e in try(each.value.fqdns, []) : {
+    pattern = try(e, null)
+  }]
+}
+
+resource "sdwan_policy_object_security_geolocation_list" "policy_object_security_geolocation_list" {
+  for_each           = { for p in try(local.feature_profiles.policy_object_profile.security_geo_location_lists, {}) : p.name => p }
+  name               = each.value.name
+  description        = null # not supported in the UI
+  feature_profile_id = sdwan_policy_object_feature_profile.policy_object_feature_profile[0].id
+  entries = flatten([
+    [for continent_code in try(each.value.continent_codes, []) : {
+      continent = continent_code
+    }],
+    [for country_code in try(each.value.country_codes, []) : {
+      country = country_code
+    }]
+  ])
+}
+
+resource "sdwan_policy_object_security_ips_signature" "policy_object_security_ips_signature" {
+  for_each           = { for p in try(local.feature_profiles.policy_object_profile.security_ips_signature_lists, {}) : p.name => p }
+  name               = each.value.name
+  feature_profile_id = sdwan_policy_object_feature_profile.policy_object_feature_profile[0].id
+  entries = [for e in try(each.value.entries, []) : {
+    generator_id = tostring(e.generator_id)
+    signature_id = tostring(e.signature_id)
+  }]
+}
+
+resource "sdwan_policy_object_security_local_application_list" "policy_object_security_local_application_list" {
+  for_each           = { for p in try(local.feature_profiles.policy_object_profile.security_local_application_lists, {}) : p.name => p }
+  name               = each.value.name
+  feature_profile_id = sdwan_policy_object_feature_profile.policy_object_feature_profile[0].id
+  entries = [for e in concat([for app in try(each.value.applications, []) : { "application" : app }], [for fam in try(each.value.application_families, []) : { "application_family" : fam }]) : {
+    app        = try(e.application, null)
+    app_family = try(e.application_family, null)
+  }]
+}
+
+resource "sdwan_policy_object_security_port_list" "policy_object_security_port_list" {
+  for_each           = { for p in try(local.feature_profiles.policy_object_profile.security_port_lists, {}) : p.name => p }
+  name               = each.value.name
+  description        = null # not supported in the UI
+  feature_profile_id = sdwan_policy_object_feature_profile.policy_object_feature_profile[0].id
+  entries = [for port_value in try(each.value.ports, []) : {
+    port = port_value
+  }]
+}
+
+resource "sdwan_policy_object_security_protocol_list" "policy_object_security_protocol_list" {
+  for_each           = { for p in try(local.feature_profiles.policy_object_profile.security_protocol_lists, {}) : p.name => p }
+  name               = each.value.name
+  description        = null # not supported in the UI
+  feature_profile_id = sdwan_policy_object_feature_profile.policy_object_feature_profile[0].id
+  entries = [for protocol in try(each.value.protocols, []) : {
+    protocol_name = protocol
+  }]
+}
+
 resource "sdwan_policy_object_sla_class_list" "policy_object_sla_class_list" {
   for_each           = { for p in try(local.feature_profiles.policy_object_profile.sla_classes, {}) : p.name => p }
   name               = each.value.name
@@ -172,17 +272,16 @@ resource "sdwan_policy_object_tloc_list" "policy_object_tloc_list" {
   }]
 }
 
-resource "sdwan_policy_object_preferred_color_group" "policy_object_preferred_color_group" {
-  for_each           = { for p in try(local.feature_profiles.policy_object_profile.preferred_color_groups, {}) : p.name => p }
-  name               = each.value.name
-  description        = try(each.value.description, null)
-  feature_profile_id = sdwan_policy_object_feature_profile.policy_object_feature_profile[0].id
-  entries = [{
-    primary_color_preference   = each.value.primary_colors
-    primary_path_preference    = try(each.value.primary_path_preference, null)
-    secondary_color_preference = try(each.value.secondary_colors, null)
-    secondary_path_preference  = try(each.value.secondary_path_preference, null)
-    tertiary_color_preference  = try(each.value.tertiary_colors, null)
-    tertiary_path_preference   = try(each.value.tertiary_path_preference, null)
-  }]
+resource "sdwan_policy_object_unified_advanced_malware_protection" "policy_object_unified_advanced_malware_protection" {
+  for_each                      = { for p in try(local.feature_profiles.policy_object_profile.security_advanced_malware_protection_profiles, {}) : p.name => p }
+  name                          = each.value.name
+  description                   = null # not supported in the UI
+  feature_profile_id            = sdwan_policy_object_feature_profile.policy_object_feature_profile[0].id
+  alert_log_level               = each.value.alert_log_level
+  amp_cloud_region              = each.value.amp_cloud_region
+  amp_cloud_region_est_server   = each.value.amp_cloud_region
+  file_analysis                 = each.value.file_analysis
+  file_analysis_alert_log_level = try(each.value.file_analysis_alert_log_level, null)
+  file_analysis_cloud_region    = try(each.value.tg_cloud_region, null)
+  file_analysis_file_types      = try(each.value.file_analysis_file_types, null)
 }